knowledge/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain-core
langchain-community
langchain-openai
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-milvus
llama-index-retrievers-bm25
llama-index-llms-langchain
llama-index-embeddings-langchain
langchain-milvus
pymilvus
pydantic-settings
markdownify
scikit-learn
jieba
llmlingua
unstructured
markdown
python-multipart
ragas
elasticsearch
